+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PIN protocol changes:
   - Extract PIN protocol implementation into separate module ([#62][])
   - Implement PIN protocol 2 ([#63][])
+  - Implement PIN token permissions ([#63][])
 
 [#26]: https://github.com/solokeys/fido-authenticator/issues/26
 [#28]: https://github.com/solokeys/fido-authenticator/issues/28